Added connection to database for backtest results Optimised download scripts Optimised symbol options

+++ api_v2/management/commands/import_binance_data.py
@@ -1,5 +1,9 @@
-from datetime import datetime
+import calendar
+from datetime import datetime, timedelta
+from itertools import permutations
+from pathlib import Path
 
+import numpy as np
 import pandas as pd
 import requests
 from django.conf import settings
@@ -7,6 +11,8 @@ from django.core.management.base import BaseCommand
 from django.db import connection
 
 from api_v2.models import KLine
+
+BINANCE_DATA_DIR: Path = settings.BINANCE_DATA_DIR
 
 
 def get_duration(start_time: datetime):
@@ -23,8 +29,18 @@ class Command(BaseCommand):
 		parser.add_argument('--drop_all', action='store_true')
 
 	def handle(self, *args, **kwargs):
-		drop_all = kwargs['drop_all']
+		tokens = settings.KLINE_TOKENS
+		import_intervals = settings.KLINE_INTERVALS
 
+		if tokens is None:
+			raise ValueError('KLINE_TOKENS is unspecified')
+
+		if import_intervals is None:
+			raise ValueError('KLINE_INERVALS is unspecified')
+
+		import_symbols = [f'{a}{b}' for a, b in permutations(tokens, 2)]
+
+		drop_all = kwargs['drop_all']
 		if drop_all:
 			msg = '\n'.join(
 				[
@@ -65,7 +81,7 @@ class Command(BaseCommand):
 			cursor.close()
 
 		self.stdout.write(self.style.SUCCESS('Pairing tokens...'))
-		symbol_folders = list(sorted(settings.BINANCE_DATA_DIR.glob('*')))
+		symbol_folders = list(sorted(BINANCE_DATA_DIR.glob('*')))
 		symbols = [folder.stem for folder in symbol_folders]
 		symbol_maps = requests.get(
 			'https://api.binance.com/api/v3/exchangeInfo?symbols=["' + '","'.join(symbols) + '"]'
@@ -82,7 +98,7 @@ class Command(BaseCommand):
 		data = []
 		for symbol_folder in symbol_folders:
 			symbol = symbol_folder.stem
-			if symbol not in settings.KLINE_SYMBOLS:
+			if symbol not in import_symbols:
 				self.stdout.write(
 					self.style.SUCCESS(f'Skipping {symbol} (Not included in ENV)... [{get_duration(start_time):d}s]')
 				)
@@ -94,13 +110,15 @@ class Command(BaseCommand):
 			to_token = symbol_maps[symbol]['to']
 			for interval_folder in sorted(symbol_folder.glob('*')):
 				interval = interval_folder.stem
-				if interval not in settings.KLINE_INTERVALS:
+				if interval not in import_intervals:
 					continue
 
 				self.stdout.write(
 					self.style.SUCCESS(f'Importing {symbol} [{interval}]... [{get_duration(start_time):d}s]')
 				)
 
+				interval_int = settings.INTERVAL_MAP[interval] * 60 * 1000
+				rows_per_day = round(60 * 24 / settings.INTERVAL_MAP[interval])
 				for file in sorted(interval_folder.glob('*')):
 					self.stdout.write(
 						self.style.SUCCESS(
@@ -131,6 +149,25 @@ class Command(BaseCommand):
 					df = df.iloc[:, 0:6]
 					df.columns = ['open_time', 'open', 'high', 'low', 'close', 'volume']
 
+					if df.iloc[0, 0] > 100000000000000:
+						df['open_time'] = df['open_time'] / 1000
+					elif df.iloc[0, 0] < 100000000000:
+						df['open_time'] = df['open_time'] * 1000
+
+					df_start_time = datetime.fromtimestamp(df.iloc[0, 0] / 1000)
+					num_days = calendar.monthrange(df_start_time.year, df_start_time.month)[1]
+					df_start_time = datetime(year=df_start_time.year, month=df_start_time.month, day=1)
+					df_end_time = df_start_time + timedelta(days=num_days)
+					start_timestamp = int(df_start_time.timestamp() * 1000)
+					end_timestamp = int(df_end_time.timestamp() * 1000)
+
+					all_time = list(range(start_timestamp, end_timestamp, interval_int))
+					full_df = pd.DataFrame({'open_time': all_time})
+					full_df = full_df.merge(df, how='left', on='open_time')
+					full_df = full_df.replace(np.nan, None)
+
+					assert len(full_df) == rows_per_day * num_days
+
 					file_data = [
 						KLine(
 							**row.to_dict(),
@@ -141,7 +178,7 @@ class Command(BaseCommand):
 							year=year,
 							month=month,
 						)
-						for _, row in df.iterrows()
+						for _, row in full_df.iterrows()
 					]
 					data.extend(file_data)
